Move _Py_DebugOffsets to its own header file.
